backend/requirements.txt
# ─── Web Framework ───────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ─── Database ────────────────────────────────
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# ─── Auth & Security ─────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0

# ─── MQTT ────────────────────────────────────
paho-mqtt>=2.0.0

# ─── ML / Data Science ───────────────────────
scikit-learn>=1.4.0
xgboost>=2.0.0
prophet>=1.1.5
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.3.0

# ─── Dashboard ───────────────────────────────
streamlit>=1.35.0
plotly>=5.22.0

# ─── Monitoring ──────────────────────────────
prometheus-fastapi-instrumentator>=7.0.0

# ─── Utilities ───────────────────────────────
httpx>=0.27.0
python-dotenv>=1.0.0
structlog>=24.0.0
apscheduler>=3.10.0

# ─── Testing ─────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
httpx>=0.27.0   # async test client for FastAPI
